+++ dynamorm/types/_schematics.py
@@ -37,9 +37,5 @@ class Schema(SchematicsModel, DynamORMSchema):
             return inst.to_primitive()
 
     @staticmethod
-    def base_field_type():
-        return types.BaseType
-
-    @staticmethod
     def base_schema_type():
         return SchematicsModel